backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
bilibili-api-python==16.0.0
yt-dlp==2023.11.16
openai-whisper==20231117
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1